cloud_function_trainer/requirements.txt
functions-framework==3.*
Flask==3.*
requests==2.*
python-dotenv==1.*
